refactor: consolidate project.types and projects.types into one file
