+++ apps/web/lib/utils/types-helpers.ts
@@ -1,11 +1,12 @@
-import type { ProjectTag, TMoney, TPercentage } from '~/lib/types/project.types'
+import type { TMoney, TPercentage, Tag } from '~/lib/types'
+import { getA11yColorMatch } from './color-utils'
 
 /** Helper function to validate and create project tag */
 export function createProjectTag(
 	id: string,
 	text: string,
 	color: { backgroundColor: string; textColor: string },
-): ProjectTag {
+): Tag {
 	if (!id.trim()) throw new Error('Tag ID cannot be empty')
 	if (!text.trim()) throw new Error('Tag text cannot be empty')
 	if (!color.backgroundColor.trim() || !color.textColor.trim()) {
@@ -37,4 +38,21 @@ export function createPercentage(value: number): TPercentage {
 		throw new Error('Percentage must be between 0 and 100')
 	}
 	return value as TPercentage
+}
+
+export function getTagColors(tag: Tag | string): {
+	backgroundColor: string
+	color: string
+} {
+	return typeof tag === 'string'
+		? { backgroundColor: '#E5E7EB', color: '#374151' }
+		: typeof tag.color !== 'string'
+			? {
+					backgroundColor: tag.color?.backgroundColor ?? '',
+					color: tag.color?.textColor ?? '',
+				}
+			: {
+					backgroundColor: tag.color,
+					color: getA11yColorMatch(tag.color)[1], // Use accessible text color
+				}
 }
